+++ service/controller/key/cluster.go
@@ -47,6 +47,7 @@ const (
 	TagCluster                 = "giantswarm.io/cluster"
 	TagClusterType             = "giantswarm.io/cluster-type"
 	TagClusterTypeControlPlane = "control-plane"
+	TagControlPlane            = "giantswarm.io/control-plane"
 	TagInstallation            = "giantswarm.io/installation"
 	TagMachineDeployment       = "giantswarm.io/machine-deployment"
 	TagOrganization            = "giantswarm.io/organization"
@@ -61,6 +62,11 @@ const (
 	StackTCCPI = "tccpi"
 	StackTCNP  = "tcnp"
 	StackTCNPF = "tcnpf"
+)
+
+const (
+	LifeCycleHookControlPlane = "ControlPlane"
+	LifeCycleHookNodePool     = "NodePool"
 )
 
 const (