+++ CHANGELOG.md
@@ -9,9 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Control Plane drainer controller.
+
+### Changed
+
+- Drop CRD management to not ensure CRDs in operators anymore.
+
 ### Fixed
 
 - Fix aws operator policy for latest node pools version.
+- Make encryption key lookup graceful during cluster creation.
+
 
 
 ## [8.2.3] 2020-04-06